clean code, add functionality, fix functionality
- clean imports, comments
- modify and add post and comment remove functionality
- fix store mobx not allowed changing state without an action error

+++ src/lib/Votes.js
@@ -4,7 +4,6 @@ import * as Users from './Users';
 import * as Posts from './Posts';
 import * as Comments from './Comments';
 import { idSeperator } from './utils';
-import { remove } from 'mobx';
 
 export const getVoteRefByVoteId = (voteId) => {
   return firestoreService.getDocumentRef('votes', voteId);
@@ -61,19 +60,29 @@ export const addVote = async (collectionName, documentId, voterId, type) => {
   }
 };
 
-export const removeVote = async (voterId, voteId) => {
-  const batch = writeBatch(firestoreService.firestore);
+export const removeVote = async (voterId, voteId, optionalBatch) => {
+  const batch = optionalBatch ?? writeBatch(firestoreService.firestore);
   const docRef = getVoteRefByVoteId(voteId);
   const vote = await getVoteById(voteId);
   const [collectionName, documentId] = idSeperator(voteId);
 
+  if (!vote) {
+    console.info(
+      'removeVote() vote from getVoteById() returned null. Aborting',
+    );
+    return;
+  }
+
   // remove doc from collection
+
   firestoreService.deleteToBatch(docRef, batch);
 
   // remove vote id from user
+
   Users.removeVoteId(voterId, voteId, batch);
 
   // increment/decrement votes object
+
   const value = vote.type === 'upvote' ? 1 : -1;
 
   if (collectionName === 'posts') {
@@ -88,11 +97,8 @@ export const removeVote = async (voterId, voteId) => {
       vote.id,
     );
   }
-
-  try {
+  if (!optionalBatch) {
     batch.commit();
-  } catch (error) {
-    console.error(error);
   }
 };
 
@@ -103,11 +109,15 @@ export const invertVote = async (voteId) => {
   const objToUpdate = { type: vote.type === 'upvote' ? 'downvote' : 'upvote' };
 
   // update doc in collection
+
   firestoreService.updateToBatch(docRef, objToUpdate, batch);
 
   // increment/decrement votes object
+
   const [collectionName, documentId] = idSeperator(voteId);
+
   // increment/decrement twice, once to revert the previous vote and once for the new vote
+
   const value = vote.type === 'upvote' ? -2 : 2;
 
   if (collectionName === 'posts') {
@@ -121,6 +131,7 @@ export const invertVote = async (voteId) => {
     );
   }
 
+  // execute
   try {
     batch.commit();
   } catch (error) {
@@ -132,13 +143,18 @@ export const handleVote = async (collectionName, documentId, voterId, type) => {
   const voteId = createVoteId(collectionName, documentId);
   const vote = await getVoteById(voteId);
   if (vote) {
+    console.info('handleVote() vote found. ');
+
     if (vote.type === type) {
       // remove vote
+      console.info('handleVote() removing vote. ');
       await removeVote(voterId, voteId);
     } else {
+      console.info('handleVote() inverting vote. ');
       await invertVote(voteId);
     }
   } else {
+    console.info('handleVote() vote not found. Adding new vote.');
     await addVote(collectionName, documentId, voterId, type);
   }
 };
